+++ SourceCode/Services/Extensions/PersonExtensions.cs
@@ -2,6 +2,7 @@
 public static class PersonExtensions
 {
     public static string NameCityAndCountry(this Person me) =>
+    me is null ? string.Empty :
     me.Country is null ? $"{me.Name()}, {me.CityName}" :
     $"{me.Name()}, {me.CityName}, {me.Country.EnglishName.Localized()}";
 }